Build updates: (1) Have libprocess only build a convenience library by default (so nothing gets installed); (2) Remove libev from Mesos makefiles (as well as src/python/setup.py.in); (3) Actually force build.cpp to get recompiled; (4) Added libleveldb.a directly to libmesos.la so that it wasn't included in the installed libmesos.la in 'dependency_lib'; (5) Added leveldb to python/setup.py.in (given the previous).
git-svn-id: https://svn.apache.org/repos/asf/incubator/mesos/trunk@1299823 13f79535-47bb-0310-9956-ffa450edef68

+++ src/python/native/proxy_scheduler.cpp
@@ -78,7 +78,7 @@ void ProxyScheduler::resourceOffers(SchedulerDriver* driver,
   if (list == NULL) {
     goto cleanup;
   }
-  for (int i = 0; i < offers.size(); i++) {
+  for (size_t i = 0; i < offers.size(); i++) {
     PyObject* offer = createPythonProtobuf(offers[i], "Offer");
     if (offer == NULL) {
       goto cleanup;
